LPS-62907 move package
