+++ modules/apps/product-navigation/product-navigation-control-menu-web/src/main/java/com/liferay/product/navigation/control/menu/web/ToggleControlsControlMenuEntry.java
@@ -12,25 +12,30 @@
  * details.
  */
 
-package com.liferay.control.menu.web;
+package com.liferay.product.navigation.control.menu.web;
 
-import com.liferay.control.menu.BaseJSPControlMenuEntry;
+import com.liferay.control.menu.BaseControlMenuEntry;
 import com.liferay.control.menu.ControlMenuEntry;
 import com.liferay.control.menu.constants.ControlMenuCategoryKeys;
 import com.liferay.portal.kernel.exception.PortalException;
+import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.security.permission.ActionKeys;
 import com.liferay.portal.kernel.service.permission.LayoutPermissionUtil;
+import com.liferay.portal.kernel.service.permission.PortletPermissionUtil;
 import com.liferay.portal.kernel.theme.ThemeDisplay;
+import com.liferay.portal.kernel.util.GetterUtil;
+import com.liferay.portal.kernel.util.SessionClicks;
 import com.liferay.portal.kernel.util.WebKeys;
+import com.liferay.portal.model.Group;
 import com.liferay.portal.model.Layout;
-import com.liferay.portal.model.LayoutTypeController;
 import com.liferay.portal.model.LayoutTypePortlet;
 
-import javax.servlet.ServletContext;
+import java.util.Locale;
+import java.util.Map;
+
 import javax.servlet.http.HttpServletRequest;
 
 import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Julio Camarero
@@ -38,22 +43,55 @@ import org.osgi.service.component.annotations.Reference;
 @Component(
 	immediate = true,
 	property = {
-		"control.menu.category.key=" + ControlMenuCategoryKeys.USER,
-		"service.ranking:Integer=100"
+		"control.menu.category.key=" + ControlMenuCategoryKeys.TOOLS,
+		"service.ranking:Integer=200"
 	},
 	service = ControlMenuEntry.class
 )
-public class AddContentControlMenuEntry
-	extends BaseJSPControlMenuEntry implements ControlMenuEntry {
+public class ToggleControlsControlMenuEntry
+	extends BaseControlMenuEntry implements ControlMenuEntry {
 
 	@Override
-	public String getBodyJspPath() {
-		return "/entries/add_content_body.jsp";
+	public Map<String, Object> getData(HttpServletRequest request) {
+		Map<String, Object> data = super.getData(request);
+
+		data.put("qa-id", "showControls");
+
+		return data;
 	}
 
 	@Override
-	public String getIconJspPath() {
-		return "/entries/add_content_icon.jsp";
+	public String getIconCssClass(HttpServletRequest request) {
+		String stateCss = null;
+
+		String toggleControls = GetterUtil.getString(
+			SessionClicks.get(
+				request, "com.liferay.frontend.js.web_toggleControls",
+				"visible"));
+
+		if (toggleControls.equals("visible")) {
+			stateCss = "view";
+		}
+		else {
+			stateCss = "hidden";
+		}
+
+		return stateCss;
+	}
+
+	@Override
+	public String getLabel(Locale locale) {
+		return LanguageUtil.get(locale, "edit-controls");
+	}
+
+	@Override
+	public String getLinkCssClass(HttpServletRequest request) {
+		return "toggle-controls";
+	}
+
+	@Override
+	public String getURL(HttpServletRequest request) {
+		return "javascript:;";
 	}
 
 	@Override
@@ -61,58 +99,26 @@ public class AddContentControlMenuEntry
 		ThemeDisplay themeDisplay = (ThemeDisplay)request.getAttribute(
 			WebKeys.THEME_DISPLAY);
 
-		if (themeDisplay.isStateMaximized()) {
-			return false;
-		}
-
 		Layout layout = themeDisplay.getLayout();
 
 		if (layout.isTypeControlPanel()) {
 			return false;
 		}
 
-		LayoutTypePortlet layoutTypePortlet =
-			themeDisplay.getLayoutTypePortlet();
+		Group group = layout.getGroup();
 
-		LayoutTypeController layoutTypeController =
-			layoutTypePortlet.getLayoutTypeController();
-
-		if (layoutTypeController.isFullPageDisplayable()) {
+		if (group.hasStagingGroup() && !group.isStagingGroup()) {
 			return false;
 		}
 
-		if (!hasAddContentOrApplicationPermission(themeDisplay)) {
-			return false;
-		}
-
-		if (!(hasCustomizePermission(themeDisplay) ||
-			  hasUpdateLayoutPermission(themeDisplay))) {
+		if (!(hasUpdateLayoutPermission(themeDisplay) ||
+			  hasCustomizePermission(themeDisplay) ||
+			  hasPortletConfigurationPermission(themeDisplay))) {
 
 			return false;
 		}
 
 		return super.isShow(request);
-	}
-
-	@Override
-	@Reference(
-		target = "(osgi.web.symbolicname=com.liferay.control.menu.web)",
-		unbind = "-"
-	)
-	public void setServletContext(ServletContext servletContext) {
-		super.setServletContext(servletContext);
-	}
-
-	protected boolean hasAddContentOrApplicationPermission(
-		ThemeDisplay themeDisplay) {
-
-		Layout layout = themeDisplay.getLayout();
-
-		if (layout.isLayoutPrototypeLinkActive()) {
-			return false;
-		}
-
-		return true;
 	}
 
 	protected boolean hasCustomizePermission(ThemeDisplay themeDisplay)
@@ -140,6 +146,15 @@ public class AddContentControlMenuEntry
 		}
 
 		return false;
+	}
+
+	protected boolean hasPortletConfigurationPermission(
+			ThemeDisplay themeDisplay)
+		throws PortalException {
+
+		return PortletPermissionUtil.hasConfigurationPermission(
+			themeDisplay.getPermissionChecker(), themeDisplay.getSiteGroupId(),
+			themeDisplay.getLayout(), ActionKeys.CONFIGURATION);
 	}
 
 	protected boolean hasUpdateLayoutPermission(ThemeDisplay themeDisplay)
