+++ modules/apps/product-navigation/product-navigation-control-menu-web/src/main/java/com/liferay/product/navigation/control/menu/web/constants/ControlMenuPortletKeys.java
@@ -12,17 +12,14 @@
  * details.
  */
 
-package com.liferay.control.menu.web.constants;
+package com.liferay.product.navigation.control.menu.web.constants;
 
 /**
- * @author Julio Camarero
+ * @author Eudaldo Alonso
  */
-public class ControlMenuWebKeys {
+public class ControlMenuPortletKeys {
 
-	public static final String PORTLET_BACK_URL = "PORTLET_BACK_URL";
-
-	public static final String PORTLET_DESCRIPTION = "PORTLET_DESCRIPTION";
-
-	public static final String PORTLET_TITLE = "PORTLET_TITLE";
+	public static final String CONTROL_MENU =
+		"com_liferay_control_menu_web_portlet_ControlMenuPortlet";
 
 }
