+++ modules/apps/product-navigation/product-navigation-control-menu-web/src/main/java/com/liferay/product/navigation/control/menu/web/portlet/ControlMenuViewPortletProvider.java
@@ -12,12 +12,12 @@
  * details.
  */
 
-package com.liferay.control.menu.web.portlet;
+package com.liferay.product.navigation.control.menu.web.portlet;
 
 import com.liferay.admin.kernel.util.PortalControlMenuApplicationType;
-import com.liferay.control.menu.web.constants.ControlMenuPortletKeys;
 import com.liferay.portal.kernel.portlet.BasePortletProvider;
 import com.liferay.portal.kernel.portlet.ViewPortletProvider;
+import com.liferay.product.navigation.control.menu.web.constants.ControlMenuPortletKeys;
 
 import org.osgi.service.component.annotations.Component;
